[Changes Made]
I have connected the forms to be able to see the data that has been
input on the products table. The plates on the dashboard also update
based on the total number of items, the total price low stock and out of
stock. Also the the quick stats displays the same